# Iterate whitespace for perf