--- a/test/parse.js
+++ b/test/parse.js
@@ -24,7 +24,7 @@ describe('cookie.parse(str)', function () {
   })
 
   it('should parse cookie with empty value', function () {
-    assert.deepEqual(cookie.parse('foo= ; bar='), { foo: '', bar: '' })
+    assert.deepEqual(cookie.parse('foo=; bar='), { foo: '', bar: '' })
   })
 
   it('should URL-decode values', function () {
@@ -36,12 +36,16 @@ describe('cookie.parse(str)', function () {
 
   it('should parse quoted values', function () {
     assert.deepEqual(cookie.parse('foo="bar"'), { foo: 'bar' })
+    assert.deepEqual(cookie.parse('foo=" a b c "'), { foo: ' a b c ' })
   })
 
   it('should trim whitespace around key and value', function () {
     assert.deepEqual(cookie.parse('  foo  =  "bar"  '), { foo: 'bar' })
     assert.deepEqual(cookie.parse('  foo  =  bar  ;  fizz  =  buzz  '), { foo: 'bar', fizz: 'buzz' })
-    assert.deepEqual(cookie.parse('foo=" a b c "'), { foo: ' a b c ' })
+    assert.deepEqual(cookie.parse(' foo = " a b c " '), { foo: ' a b c ' })
+    assert.deepEqual(cookie.parse(' = bar '), { '': 'bar' })
+    assert.deepEqual(cookie.parse(' foo = '), { foo: '' })
+    assert.deepEqual(cookie.parse('   =   '), { '': '' })
   })
 
   it('should return original value on escape error', function () {