@blakeembrey 2024-10-02
Undid the `=` loop for now since perf falls behind after ~10 characters, so it'd be a matter of deciding whether most cookie names are < 10 characters or not to adopt that change.